+++ Backend/PriceTag/src/main/java/com/antunmod/pricetag/repo/SectorRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 import com.antunmod.pricetag.model.database.Sector;
 
 @Service
-public interface SectorRepository extends JpaRepository<Sector, Long> {
+public interface SectorRepository extends JpaRepository<Sector, Byte> {
 
 	@Query(value = "SELECT sector_name FROM sector", nativeQuery = true)
 	List<String> getAllSectorNames();
