+++ Backend/PriceTag/src/main/java/com/antunmod/pricetag/repo/SizeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.antunmod.pricetag.model.database.Size;
 
-public interface SizeRepository extends JpaRepository<Size, Long> {
+public interface SizeRepository extends JpaRepository<Size, Byte> {
 
 	@Query(value = "SELECT size_type FROM product_size", nativeQuery = true)
 	List<String> getSizeTypes();
