+++ Backend/PriceTag/src/main/java/com/antunmod/pricetag/repo/SubcategoryProductRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 import com.antunmod.pricetag.model.database.SubcategoryProduct;
 
 @Service
-public interface SubcategoryProductRepository extends JpaRepository<SubcategoryProduct, Long> {
+public interface SubcategoryProductRepository extends JpaRepository<SubcategoryProduct, Short> {
 
 	@Query(value = "SELECT subcategory_ID FROM subcategory where subcategory_name = ?1 ", nativeQuery = true)
 	Integer getSubcategoryIdForSubcategoryName(String subcategoryName);
