+++ Backend/PriceTag/src/main/java/com/antunmod/pricetag/repo/SubcategoryRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 import com.antunmod.pricetag.model.database.Subcategory;
 
 @Service
-public interface SubcategoryRepository extends JpaRepository<Subcategory, Long> {
+public interface SubcategoryRepository extends JpaRepository<Subcategory, Short> {
 
 	@Query(value = "SELECT subcategory_name FROM category NATURAL JOIN category_subcategory NATURAL JOIN subcategory "
 			+ "WHERE category_name = ?1", nativeQuery = true)
