+++ Backend/PriceTag/src/main/java/com/antunmod/pricetag/repo/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 import com.antunmod.pricetag.model.database.User;
 
 @Service
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Short> {
 
 	// FIND_USER
 	@Query(value = "SELECT * FROM user u WHERE u.user_name LIKE ?1", nativeQuery = true)
@@ -16,6 +16,6 @@ public interface UserRepository extends JpaRepository<User, Long> {
 	@Query(value = "SELECT * FROM user u WHERE u.user_name LIKE ?1 and user_password LIKE ?2", nativeQuery = true)
 	User findByUserNameAndPassword(String username, String password);
 
-	User findByUserId(long userId);
+	User findByUserId(short userId);
 
 }
